test(framework/image,ui): rename test files to describe contents
Mechanical renames so the file names match what's actually inside —
"stress" and "r4_" were carrying review-round connotations that
don't tell future readers anything useful.

  framework/image/stress_test.go               → edge_test.go
  framework/image/fuzz_security_test.go        → fuzz_test.go
  framework/image/internal/vp8l/stress_test.go → roundtrip_test.go
  framework/image/internal/vp8l/r4_boundary_test.go → boundary_test.go

Test function prefixes follow the file rename:
  TestStress*  → TestRoundTrip*  (every test in vp8l/roundtrip_test
                                  is a pixel-equal round-trip)
  TestR4_*     → Test*           (drop the review-round prefix)

framework/ui/pipeline_image_stress_test.go (3 funcs) folded into
pipeline_image_test.go — only the XSS-escape test survived, the
duplicate-URL and 1000-source tests were t.Logf-only and the dedup
behaviour they documented was superseded by the round-3 dedup test.
