+++ framework/image/edge_test.go
