+++ framework/image/fuzz_test.go
